09-FEB-2025/CVT/defect-detection-app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
numpy==1.21.0
opencv-python-headless==4.5.3.56
Pillow==8.3.1
scipy==1.7.0
gunicorn==20.1.0
